chore(Occurrences): move closer/fetcher job classes to ProCiv namespace
